Enum list_user_pools::Error

#[non_exhaustive]
pub(crate) enum Error {

    AliasExistsException(AliasExistsException),
    CodeDeliveryFailureException(CodeDeliveryFailureException),
    CodeMismatchException(CodeMismatchException),
    ConcurrentModificationException(ConcurrentModificationException),
    DuplicateProviderException(DuplicateProviderException),
    EnableSoftwareTokenMfaException(EnableSoftwareTokenMfaException),
    ExpiredCodeException(ExpiredCodeException),
    GroupExistsException(GroupExistsException),
    InternalErrorException(InternalErrorException),
    InvalidEmailRoleAccessPolicyException(InvalidEmailRoleAccessPolicyException),
    InvalidLambdaResponseException(InvalidLambdaResponseException),
    InvalidOAuthFlowException(InvalidOAuthFlowException),
    InvalidParameterException(InvalidParameterException),
    InvalidPasswordException(InvalidPasswordException),
    InvalidSmsRoleAccessPolicyException(InvalidSmsRoleAccessPolicyException),
    InvalidSmsRoleTrustRelationshipException(InvalidSmsRoleTrustRelationshipException),
    InvalidUserPoolConfigurationException(InvalidUserPoolConfigurationException),
    LimitExceededException(LimitExceededException),
    MfaMethodNotFoundException(MfaMethodNotFoundException),
    NotAuthorizedException(NotAuthorizedException),
    PasswordResetRequiredException(PasswordResetRequiredException),
    PreconditionNotMetException(PreconditionNotMetException),
    ResourceNotFoundException(ResourceNotFoundException),
    ScopeDoesNotExistException(ScopeDoesNotExistException),
    SoftwareTokenMfaNotFoundException(SoftwareTokenMfaNotFoundException),
    TooManyFailedAttemptsException(TooManyFailedAttemptsException),
    TooManyRequestsException(TooManyRequestsException),
    UnauthorizedException(UnauthorizedException),
    UnexpectedLambdaException(UnexpectedLambdaException),
    UnsupportedIdentityProviderException(UnsupportedIdentityProviderException),
    UnsupportedOperationException(UnsupportedOperationException),
    UnsupportedTokenTypeException(UnsupportedTokenTypeException),
    UnsupportedUserStateException(UnsupportedUserStateException),
    UserImportInProgressException(UserImportInProgressException),
    UserLambdaValidationException(UserLambdaValidationException),
    UserNotConfirmedException(UserNotConfirmedException),
    UserNotFoundException(UserNotFoundException),
    UserPoolAddOnNotEnabledException(UserPoolAddOnNotEnabledException),
    UserPoolTaggingException(UserPoolTaggingException),
    UsernameExistsException(UsernameExistsException),
    Unhandled(Box<dyn Error + Send + Sync + 'static, Global>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AliasExistsException(AliasExistsException)

This exception is thrown when a user tries to confirm the account with an email or phone number that has already been supplied as an alias from a different account. This exception tells user that an account with this email or phone already exists.

Tuple Fields of AliasExistsException

0: AliasExistsException

CodeDeliveryFailureException(CodeDeliveryFailureException)

This exception is thrown when a verification code fails to deliver successfully.

Tuple Fields of CodeDeliveryFailureException

0: CodeDeliveryFailureException

CodeMismatchException(CodeMismatchException)

This exception is thrown if the provided code does not match what the server was expecting.

Tuple Fields of CodeMismatchException

0: CodeMismatchException

ConcurrentModificationException(ConcurrentModificationException)

This exception is thrown if two or more modifications are happening concurrently.

Tuple Fields of ConcurrentModificationException

0: ConcurrentModificationException

DuplicateProviderException(DuplicateProviderException)

This exception is thrown when the provider is already supported by the user pool.

Tuple Fields of DuplicateProviderException

0: DuplicateProviderException

EnableSoftwareTokenMfaException(EnableSoftwareTokenMfaException)

This exception is thrown when there is a code mismatch and the service fails to configure the software token TOTP multi-factor authentication (MFA).

Tuple Fields of EnableSoftwareTokenMfaException

0: EnableSoftwareTokenMfaException

ExpiredCodeException(ExpiredCodeException)

This exception is thrown if a code has expired.

Tuple Fields of ExpiredCodeException

0: ExpiredCodeException

GroupExistsException(GroupExistsException)

This exception is thrown when Amazon Cognito encounters a group that already exists in the user pool.

Tuple Fields of GroupExistsException

0: GroupExistsException

InternalErrorException(InternalErrorException)

This exception is thrown when Amazon Cognito encounters an internal error.

Tuple Fields of InternalErrorException

0: InternalErrorException

InvalidEmailRoleAccessPolicyException(InvalidEmailRoleAccessPolicyException)

This exception is thrown when Amazon Cognito is not allowed to use your email identity. HTTP status code: 400.

Tuple Fields of InvalidEmailRoleAccessPolicyException

0: InvalidEmailRoleAccessPolicyException

InvalidLambdaResponseException(InvalidLambdaResponseException)

This exception is thrown when the Amazon Cognito service encounters an invalid Lambda response.

Tuple Fields of InvalidLambdaResponseException

0: InvalidLambdaResponseException

InvalidOAuthFlowException(InvalidOAuthFlowException)

This exception is thrown when the specified OAuth flow is invalid.

Tuple Fields of InvalidOAuthFlowException

0: InvalidOAuthFlowException

InvalidParameterException(InvalidParameterException)

This exception is thrown when the Amazon Cognito service encounters an invalid parameter.

Tuple Fields of InvalidParameterException

0: InvalidParameterException

InvalidPasswordException(InvalidPasswordException)

This exception is thrown when the Amazon Cognito service encounters an invalid password.

Tuple Fields of InvalidPasswordException

0: InvalidPasswordException

InvalidSmsRoleAccessPolicyException(InvalidSmsRoleAccessPolicyException)

This exception is returned when the role provided for SMS configuration does not have permission to publish using Amazon SNS.

Tuple Fields of InvalidSmsRoleAccessPolicyException

0: InvalidSmsRoleAccessPolicyException

InvalidSmsRoleTrustRelationshipException(InvalidSmsRoleTrustRelationshipException)

This exception is thrown when the trust relationship is invalid for the role provided for SMS configuration. This can happen if you do not trust cognito-idp.amazonaws.com or the external ID provided in the role does not match what is provided in the SMS configuration for the user pool.

Tuple Fields of InvalidSmsRoleTrustRelationshipException

0: InvalidSmsRoleTrustRelationshipException

InvalidUserPoolConfigurationException(InvalidUserPoolConfigurationException)

This exception is thrown when the user pool configuration is invalid.

Tuple Fields of InvalidUserPoolConfigurationException

0: InvalidUserPoolConfigurationException

LimitExceededException(LimitExceededException)

This exception is thrown when a user exceeds the limit for a requested Amazon Web Services resource.

Tuple Fields of LimitExceededException

0: LimitExceededException

MfaMethodNotFoundException(MfaMethodNotFoundException)

This exception is thrown when Amazon Cognito cannot find a multi-factor authentication (MFA) method.

Tuple Fields of MfaMethodNotFoundException

0: MfaMethodNotFoundException

NotAuthorizedException(NotAuthorizedException)

This exception is thrown when a user is not authorized.

Tuple Fields of NotAuthorizedException

0: NotAuthorizedException

PasswordResetRequiredException(PasswordResetRequiredException)

This exception is thrown when a password reset is required.

Tuple Fields of PasswordResetRequiredException

0: PasswordResetRequiredException

PreconditionNotMetException(PreconditionNotMetException)

This exception is thrown when a precondition is not met.

Tuple Fields of PreconditionNotMetException

0: PreconditionNotMetException

ResourceNotFoundException(ResourceNotFoundException)

This exception is thrown when the Amazon Cognito service cannot find the requested resource.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ScopeDoesNotExistException(ScopeDoesNotExistException)

This exception is thrown when the specified scope does not exist.

Tuple Fields of ScopeDoesNotExistException

0: ScopeDoesNotExistException

SoftwareTokenMfaNotFoundException(SoftwareTokenMfaNotFoundException)

This exception is thrown when the software token TOTP multi-factor authentication (MFA) is not enabled for the user pool.

Tuple Fields of SoftwareTokenMfaNotFoundException

0: SoftwareTokenMfaNotFoundException

TooManyFailedAttemptsException(TooManyFailedAttemptsException)

This exception is thrown when the user has made too many failed attempts for a given action (e.g., sign in).

Tuple Fields of TooManyFailedAttemptsException

0: TooManyFailedAttemptsException

TooManyRequestsException(TooManyRequestsException)

This exception is thrown when the user has made too many requests for a given operation.

Tuple Fields of TooManyRequestsException

0: TooManyRequestsException

UnauthorizedException(UnauthorizedException)

This exception is thrown when the request is not authorized. This can happen due to an invalid access token in the request.

Tuple Fields of UnauthorizedException

0: UnauthorizedException

UnexpectedLambdaException(UnexpectedLambdaException)

This exception is thrown when the Amazon Cognito service encounters an unexpected exception with the Lambda service.

Tuple Fields of UnexpectedLambdaException

0: UnexpectedLambdaException

UnsupportedIdentityProviderException(UnsupportedIdentityProviderException)

This exception is thrown when the specified identifier is not supported.

Tuple Fields of UnsupportedIdentityProviderException

0: UnsupportedIdentityProviderException

UnsupportedOperationException(UnsupportedOperationException)

This exception is thrown when you attempt to perform an operation that is not enabled for the user pool client.

Tuple Fields of UnsupportedOperationException

0: UnsupportedOperationException

UnsupportedTokenTypeException(UnsupportedTokenTypeException)

This exception is thrown when an unsupported token is passed to an operation.

Tuple Fields of UnsupportedTokenTypeException

0: UnsupportedTokenTypeException

UnsupportedUserStateException(UnsupportedUserStateException)

The request failed because the user is in an unsupported state.

Tuple Fields of UnsupportedUserStateException

0: UnsupportedUserStateException

UserImportInProgressException(UserImportInProgressException)

This exception is thrown when you are trying to modify a user pool while a user import job is in progress for that pool.

Tuple Fields of UserImportInProgressException

0: UserImportInProgressException

UserLambdaValidationException(UserLambdaValidationException)

This exception is thrown when the Amazon Cognito service encounters a user validation exception with the Lambda service.

Tuple Fields of UserLambdaValidationException

0: UserLambdaValidationException

UserNotConfirmedException(UserNotConfirmedException)

This exception is thrown when a user is not confirmed successfully.

Tuple Fields of UserNotConfirmedException

0: UserNotConfirmedException

UserNotFoundException(UserNotFoundException)

This exception is thrown when a user is not found.

Tuple Fields of UserNotFoundException

0: UserNotFoundException

UserPoolAddOnNotEnabledException(UserPoolAddOnNotEnabledException)

This exception is thrown when user pool add-ons are not enabled.

Tuple Fields of UserPoolAddOnNotEnabledException

0: UserPoolAddOnNotEnabledException

UserPoolTaggingException(UserPoolTaggingException)

This exception is thrown when a user pool tag cannot be set or updated.

Tuple Fields of UserPoolTaggingException

0: UserPoolTaggingException

UsernameExistsException(UsernameExistsException)

This exception is thrown when Amazon Cognito encounters a user name that already exists in the user pool.

Tuple Fields of UsernameExistsException

0: UsernameExistsException

Unhandled(Box<dyn Error + Send + Sync + 'static, Global>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static, Global>