Struct list_user_pools::Client

pub(crate) struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> {
    handle: Arc<Handle<C, M, R>>,
}

An ergonomic service client for AWSCognitoIdentityProviderService.

This client allows ergonomic access to a AWSCognitoIdentityProviderService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Fields
handle: Arc<Handle<C, M, R>>
Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Client<C, M, R>

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn add_custom_attributes(&self) -> AddCustomAttributes<C, M, R>

Constructs a fluent builder for the AddCustomAttributes operation.

See AddCustomAttributes for more information about the operation and its arguments.

pub fn admin_add_user_to_group(&self) -> AdminAddUserToGroup<C, M, R>

Constructs a fluent builder for the AdminAddUserToGroup operation.

See AdminAddUserToGroup for more information about the operation and its arguments.

pub fn admin_confirm_sign_up(&self) -> AdminConfirmSignUp<C, M, R>

Constructs a fluent builder for the AdminConfirmSignUp operation.

See AdminConfirmSignUp for more information about the operation and its arguments.

pub fn admin_create_user(&self) -> AdminCreateUser<C, M, R>

Constructs a fluent builder for the AdminCreateUser operation.

See AdminCreateUser for more information about the operation and its arguments.

pub fn admin_delete_user(&self) -> AdminDeleteUser<C, M, R>

Constructs a fluent builder for the AdminDeleteUser operation.

See AdminDeleteUser for more information about the operation and its arguments.

pub fn admin_delete_user_attributes(&self) -> AdminDeleteUserAttributes<C, M, R>

Constructs a fluent builder for the AdminDeleteUserAttributes operation.

See AdminDeleteUserAttributes for more information about the operation and its arguments.

pub fn admin_disable_provider_for_user(
&self
) -> AdminDisableProviderForUser<C, M, R>

Constructs a fluent builder for the AdminDisableProviderForUser operation.

See AdminDisableProviderForUser for more information about the operation and its arguments.

pub fn admin_disable_user(&self) -> AdminDisableUser<C, M, R>

Constructs a fluent builder for the AdminDisableUser operation.

See AdminDisableUser for more information about the operation and its arguments.

pub fn admin_enable_user(&self) -> AdminEnableUser<C, M, R>

Constructs a fluent builder for the AdminEnableUser operation.

See AdminEnableUser for more information about the operation and its arguments.

pub fn admin_forget_device(&self) -> AdminForgetDevice<C, M, R>

Constructs a fluent builder for the AdminForgetDevice operation.

See AdminForgetDevice for more information about the operation and its arguments.

pub fn admin_get_device(&self) -> AdminGetDevice<C, M, R>

Constructs a fluent builder for the AdminGetDevice operation.

See AdminGetDevice for more information about the operation and its arguments.

pub fn admin_get_user(&self) -> AdminGetUser<C, M, R>

Constructs a fluent builder for the AdminGetUser operation.

See AdminGetUser for more information about the operation and its arguments.

pub fn admin_initiate_auth(&self) -> AdminInitiateAuth<C, M, R>

Constructs a fluent builder for the AdminInitiateAuth operation.

See AdminInitiateAuth for more information about the operation and its arguments.

pub fn admin_link_provider_for_user(&self) -> AdminLinkProviderForUser<C, M, R>

Constructs a fluent builder for the AdminLinkProviderForUser operation.

See AdminLinkProviderForUser for more information about the operation and its arguments.

pub fn admin_list_devices(&self) -> AdminListDevices<C, M, R>

Constructs a fluent builder for the AdminListDevices operation.

See AdminListDevices for more information about the operation and its arguments.

pub fn admin_list_groups_for_user(&self) -> AdminListGroupsForUser<C, M, R>

Constructs a fluent builder for the AdminListGroupsForUser operation.

See AdminListGroupsForUser for more information about the operation and its arguments.

pub fn admin_list_user_auth_events(&self) -> AdminListUserAuthEvents<C, M, R>

Constructs a fluent builder for the AdminListUserAuthEvents operation.

See AdminListUserAuthEvents for more information about the operation and its arguments.

pub fn admin_remove_user_from_group(&self) -> AdminRemoveUserFromGroup<C, M, R>

Constructs a fluent builder for the AdminRemoveUserFromGroup operation.

See AdminRemoveUserFromGroup for more information about the operation and its arguments.

pub fn admin_reset_user_password(&self) -> AdminResetUserPassword<C, M, R>

Constructs a fluent builder for the AdminResetUserPassword operation.

See AdminResetUserPassword for more information about the operation and its arguments.

pub fn admin_respond_to_auth_challenge(
&self
) -> AdminRespondToAuthChallenge<C, M, R>

Constructs a fluent builder for the AdminRespondToAuthChallenge operation.

See AdminRespondToAuthChallenge for more information about the operation and its arguments.

pub fn admin_set_user_mfa_preference(
&self
) -> AdminSetUserMFAPreference<C, M, R>

Constructs a fluent builder for the AdminSetUserMFAPreference operation.

See AdminSetUserMFAPreference for more information about the operation and its arguments.

pub fn admin_set_user_password(&self) -> AdminSetUserPassword<C, M, R>

Constructs a fluent builder for the AdminSetUserPassword operation.

See AdminSetUserPassword for more information about the operation and its arguments.

pub fn admin_set_user_settings(&self) -> AdminSetUserSettings<C, M, R>

Constructs a fluent builder for the AdminSetUserSettings operation.

See AdminSetUserSettings for more information about the operation and its arguments.

pub fn admin_update_auth_event_feedback(
&self
) -> AdminUpdateAuthEventFeedback<C, M, R>

Constructs a fluent builder for the AdminUpdateAuthEventFeedback operation.

See AdminUpdateAuthEventFeedback for more information about the operation and its arguments.

pub fn admin_update_device_status(&self) -> AdminUpdateDeviceStatus<C, M, R>

Constructs a fluent builder for the AdminUpdateDeviceStatus operation.

See AdminUpdateDeviceStatus for more information about the operation and its arguments.

pub fn admin_update_user_attributes(&self) -> AdminUpdateUserAttributes<C, M, R>

Constructs a fluent builder for the AdminUpdateUserAttributes operation.

See AdminUpdateUserAttributes for more information about the operation and its arguments.

pub fn admin_user_global_sign_out(&self) -> AdminUserGlobalSignOut<C, M, R>

Constructs a fluent builder for the AdminUserGlobalSignOut operation.

See AdminUserGlobalSignOut for more information about the operation and its arguments.

pub fn associate_software_token(&self) -> AssociateSoftwareToken<C, M, R>

Constructs a fluent builder for the AssociateSoftwareToken operation.

See AssociateSoftwareToken for more information about the operation and its arguments.

pub fn change_password(&self) -> ChangePassword<C, M, R>

Constructs a fluent builder for the ChangePassword operation.

See ChangePassword for more information about the operation and its arguments.

pub fn confirm_device(&self) -> ConfirmDevice<C, M, R>

Constructs a fluent builder for the ConfirmDevice operation.

See ConfirmDevice for more information about the operation and its arguments.

pub fn confirm_forgot_password(&self) -> ConfirmForgotPassword<C, M, R>

Constructs a fluent builder for the ConfirmForgotPassword operation.

See ConfirmForgotPassword for more information about the operation and its arguments.

pub fn confirm_sign_up(&self) -> ConfirmSignUp<C, M, R>

Constructs a fluent builder for the ConfirmSignUp operation.

See ConfirmSignUp for more information about the operation and its arguments.

pub fn create_group(&self) -> CreateGroup<C, M, R>

Constructs a fluent builder for the CreateGroup operation.

See CreateGroup for more information about the operation and its arguments.

pub fn create_identity_provider(&self) -> CreateIdentityProvider<C, M, R>

Constructs a fluent builder for the CreateIdentityProvider operation.

See CreateIdentityProvider for more information about the operation and its arguments.

pub fn create_resource_server(&self) -> CreateResourceServer<C, M, R>

Constructs a fluent builder for the CreateResourceServer operation.

See CreateResourceServer for more information about the operation and its arguments.

pub fn create_user_import_job(&self) -> CreateUserImportJob<C, M, R>

Constructs a fluent builder for the CreateUserImportJob operation.

See CreateUserImportJob for more information about the operation and its arguments.

pub fn create_user_pool(&self) -> CreateUserPool<C, M, R>

Constructs a fluent builder for the CreateUserPool operation.

See CreateUserPool for more information about the operation and its arguments.

pub fn create_user_pool_client(&self) -> CreateUserPoolClient<C, M, R>

Constructs a fluent builder for the CreateUserPoolClient operation.

See CreateUserPoolClient for more information about the operation and its arguments.

pub fn create_user_pool_domain(&self) -> CreateUserPoolDomain<C, M, R>

Constructs a fluent builder for the CreateUserPoolDomain operation.

See CreateUserPoolDomain for more information about the operation and its arguments.

pub fn delete_group(&self) -> DeleteGroup<C, M, R>

Constructs a fluent builder for the DeleteGroup operation.

See DeleteGroup for more information about the operation and its arguments.

pub fn delete_identity_provider(&self) -> DeleteIdentityProvider<C, M, R>

Constructs a fluent builder for the DeleteIdentityProvider operation.

See DeleteIdentityProvider for more information about the operation and its arguments.

pub fn delete_resource_server(&self) -> DeleteResourceServer<C, M, R>

Constructs a fluent builder for the DeleteResourceServer operation.

See DeleteResourceServer for more information about the operation and its arguments.

pub fn delete_user(&self) -> DeleteUser<C, M, R>

Constructs a fluent builder for the DeleteUser operation.

See DeleteUser for more information about the operation and its arguments.

pub fn delete_user_attributes(&self) -> DeleteUserAttributes<C, M, R>

Constructs a fluent builder for the DeleteUserAttributes operation.

See DeleteUserAttributes for more information about the operation and its arguments.

pub fn delete_user_pool(&self) -> DeleteUserPool<C, M, R>

Constructs a fluent builder for the DeleteUserPool operation.

See DeleteUserPool for more information about the operation and its arguments.

pub fn delete_user_pool_client(&self) -> DeleteUserPoolClient<C, M, R>

Constructs a fluent builder for the DeleteUserPoolClient operation.

See DeleteUserPoolClient for more information about the operation and its arguments.

pub fn delete_user_pool_domain(&self) -> DeleteUserPoolDomain<C, M, R>

Constructs a fluent builder for the DeleteUserPoolDomain operation.

See DeleteUserPoolDomain for more information about the operation and its arguments.

pub fn describe_identity_provider(&self) -> DescribeIdentityProvider<C, M, R>

Constructs a fluent builder for the DescribeIdentityProvider operation.

See DescribeIdentityProvider for more information about the operation and its arguments.

pub fn describe_resource_server(&self) -> DescribeResourceServer<C, M, R>

Constructs a fluent builder for the DescribeResourceServer operation.

See DescribeResourceServer for more information about the operation and its arguments.

pub fn describe_risk_configuration(&self) -> DescribeRiskConfiguration<C, M, R>

Constructs a fluent builder for the DescribeRiskConfiguration operation.

See DescribeRiskConfiguration for more information about the operation and its arguments.

pub fn describe_user_import_job(&self) -> DescribeUserImportJob<C, M, R>

Constructs a fluent builder for the DescribeUserImportJob operation.

See DescribeUserImportJob for more information about the operation and its arguments.

pub fn describe_user_pool(&self) -> DescribeUserPool<C, M, R>

Constructs a fluent builder for the DescribeUserPool operation.

See DescribeUserPool for more information about the operation and its arguments.

pub fn describe_user_pool_client(&self) -> DescribeUserPoolClient<C, M, R>

Constructs a fluent builder for the DescribeUserPoolClient operation.

See DescribeUserPoolClient for more information about the operation and its arguments.

pub fn describe_user_pool_domain(&self) -> DescribeUserPoolDomain<C, M, R>

Constructs a fluent builder for the DescribeUserPoolDomain operation.

See DescribeUserPoolDomain for more information about the operation and its arguments.

pub fn forget_device(&self) -> ForgetDevice<C, M, R>

Constructs a fluent builder for the ForgetDevice operation.

See ForgetDevice for more information about the operation and its arguments.

pub fn forgot_password(&self) -> ForgotPassword<C, M, R>

Constructs a fluent builder for the ForgotPassword operation.

See ForgotPassword for more information about the operation and its arguments.

pub fn get_csv_header(&self) -> GetCSVHeader<C, M, R>

Constructs a fluent builder for the GetCSVHeader operation.

See GetCSVHeader for more information about the operation and its arguments.

pub fn get_device(&self) -> GetDevice<C, M, R>

Constructs a fluent builder for the GetDevice operation.

See GetDevice for more information about the operation and its arguments.

pub fn get_group(&self) -> GetGroup<C, M, R>

Constructs a fluent builder for the GetGroup operation.

See GetGroup for more information about the operation and its arguments.

pub fn get_identity_provider_by_identifier(
&self
) -> GetIdentityProviderByIdentifier<C, M, R>

Constructs a fluent builder for the GetIdentityProviderByIdentifier operation.

See GetIdentityProviderByIdentifier for more information about the operation and its arguments.

pub fn get_signing_certificate(&self) -> GetSigningCertificate<C, M, R>

Constructs a fluent builder for the GetSigningCertificate operation.

See GetSigningCertificate for more information about the operation and its arguments.

pub fn get_ui_customization(&self) -> GetUICustomization<C, M, R>

Constructs a fluent builder for the GetUICustomization operation.

See GetUICustomization for more information about the operation and its arguments.

pub fn get_user(&self) -> GetUser<C, M, R>

Constructs a fluent builder for the GetUser operation.

See GetUser for more information about the operation and its arguments.

pub fn get_user_attribute_verification_code(
&self
) -> GetUserAttributeVerificationCode<C, M, R>

Constructs a fluent builder for the GetUserAttributeVerificationCode operation.

See GetUserAttributeVerificationCode for more information about the operation and its arguments.

pub fn get_user_pool_mfa_config(&self) -> GetUserPoolMfaConfig<C, M, R>

Constructs a fluent builder for the GetUserPoolMfaConfig operation.

See GetUserPoolMfaConfig for more information about the operation and its arguments.

pub fn global_sign_out(&self) -> GlobalSignOut<C, M, R>

Constructs a fluent builder for the GlobalSignOut operation.

See GlobalSignOut for more information about the operation and its arguments.

pub fn initiate_auth(&self) -> InitiateAuth<C, M, R>

Constructs a fluent builder for the InitiateAuth operation.

See InitiateAuth for more information about the operation and its arguments.

pub fn list_devices(&self) -> ListDevices<C, M, R>

Constructs a fluent builder for the ListDevices operation.

See ListDevices for more information about the operation and its arguments.

pub fn list_groups(&self) -> ListGroups<C, M, R>

Constructs a fluent builder for the ListGroups operation.

See ListGroups for more information about the operation and its arguments.

pub fn list_identity_providers(&self) -> ListIdentityProviders<C, M, R>

Constructs a fluent builder for the ListIdentityProviders operation.

See ListIdentityProviders for more information about the operation and its arguments.

pub fn list_resource_servers(&self) -> ListResourceServers<C, M, R>

Constructs a fluent builder for the ListResourceServers operation.

See ListResourceServers for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_user_import_jobs(&self) -> ListUserImportJobs<C, M, R>

Constructs a fluent builder for the ListUserImportJobs operation.

See ListUserImportJobs for more information about the operation and its arguments.

pub fn list_user_pool_clients(&self) -> ListUserPoolClients<C, M, R>

Constructs a fluent builder for the ListUserPoolClients operation.

See ListUserPoolClients for more information about the operation and its arguments.

pub fn list_user_pools(&self) -> ListUserPools<C, M, R>

Constructs a fluent builder for the ListUserPools operation.

See ListUserPools for more information about the operation and its arguments.

pub fn list_users(&self) -> ListUsers<C, M, R>

Constructs a fluent builder for the ListUsers operation.

See ListUsers for more information about the operation and its arguments.

pub fn list_users_in_group(&self) -> ListUsersInGroup<C, M, R>

Constructs a fluent builder for the ListUsersInGroup operation.

See ListUsersInGroup for more information about the operation and its arguments.

pub fn resend_confirmation_code(&self) -> ResendConfirmationCode<C, M, R>

Constructs a fluent builder for the ResendConfirmationCode operation.

See ResendConfirmationCode for more information about the operation and its arguments.

pub fn respond_to_auth_challenge(&self) -> RespondToAuthChallenge<C, M, R>

Constructs a fluent builder for the RespondToAuthChallenge operation.

See RespondToAuthChallenge for more information about the operation and its arguments.

pub fn revoke_token(&self) -> RevokeToken<C, M, R>

Constructs a fluent builder for the RevokeToken operation.

See RevokeToken for more information about the operation and its arguments.

pub fn set_risk_configuration(&self) -> SetRiskConfiguration<C, M, R>

Constructs a fluent builder for the SetRiskConfiguration operation.

See SetRiskConfiguration for more information about the operation and its arguments.

pub fn set_ui_customization(&self) -> SetUICustomization<C, M, R>

Constructs a fluent builder for the SetUICustomization operation.

See SetUICustomization for more information about the operation and its arguments.

pub fn set_user_mfa_preference(&self) -> SetUserMFAPreference<C, M, R>

Constructs a fluent builder for the SetUserMFAPreference operation.

See SetUserMFAPreference for more information about the operation and its arguments.

pub fn set_user_pool_mfa_config(&self) -> SetUserPoolMfaConfig<C, M, R>

Constructs a fluent builder for the SetUserPoolMfaConfig operation.

See SetUserPoolMfaConfig for more information about the operation and its arguments.

pub fn set_user_settings(&self) -> SetUserSettings<C, M, R>

Constructs a fluent builder for the SetUserSettings operation.

See SetUserSettings for more information about the operation and its arguments.

pub fn sign_up(&self) -> SignUp<C, M, R>

Constructs a fluent builder for the SignUp operation.

See SignUp for more information about the operation and its arguments.

pub fn start_user_import_job(&self) -> StartUserImportJob<C, M, R>

Constructs a fluent builder for the StartUserImportJob operation.

See StartUserImportJob for more information about the operation and its arguments.

pub fn stop_user_import_job(&self) -> StopUserImportJob<C, M, R>

Constructs a fluent builder for the StopUserImportJob operation.

See StopUserImportJob for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_auth_event_feedback(&self) -> UpdateAuthEventFeedback<C, M, R>

Constructs a fluent builder for the UpdateAuthEventFeedback operation.

See UpdateAuthEventFeedback for more information about the operation and its arguments.

pub fn update_device_status(&self) -> UpdateDeviceStatus<C, M, R>

Constructs a fluent builder for the UpdateDeviceStatus operation.

See UpdateDeviceStatus for more information about the operation and its arguments.

pub fn update_group(&self) -> UpdateGroup<C, M, R>

Constructs a fluent builder for the UpdateGroup operation.

See UpdateGroup for more information about the operation and its arguments.

pub fn update_identity_provider(&self) -> UpdateIdentityProvider<C, M, R>

Constructs a fluent builder for the UpdateIdentityProvider operation.

See UpdateIdentityProvider for more information about the operation and its arguments.

pub fn update_resource_server(&self) -> UpdateResourceServer<C, M, R>

Constructs a fluent builder for the UpdateResourceServer operation.

See UpdateResourceServer for more information about the operation and its arguments.

pub fn update_user_attributes(&self) -> UpdateUserAttributes<C, M, R>

Constructs a fluent builder for the UpdateUserAttributes operation.

See UpdateUserAttributes for more information about the operation and its arguments.

pub fn update_user_pool(&self) -> UpdateUserPool<C, M, R>

Constructs a fluent builder for the UpdateUserPool operation.

See UpdateUserPool for more information about the operation and its arguments.

pub fn update_user_pool_client(&self) -> UpdateUserPoolClient<C, M, R>

Constructs a fluent builder for the UpdateUserPoolClient operation.

See UpdateUserPoolClient for more information about the operation and its arguments.

pub fn update_user_pool_domain(&self) -> UpdateUserPoolDomain<C, M, R>

Constructs a fluent builder for the UpdateUserPoolDomain operation.

See UpdateUserPoolDomain for more information about the operation and its arguments.

pub fn verify_software_token(&self) -> VerifySoftwareToken<C, M, R>

Constructs a fluent builder for the VerifySoftwareToken operation.

See VerifySoftwareToken for more information about the operation and its arguments.

pub fn verify_user_attribute(&self) -> VerifyUserAttribute<C, M, R>

Constructs a fluent builder for the VerifyUserAttribute operation.

See VerifyUserAttribute for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(
conf: Config,
conn: C
) -> Client<C, AwsMiddleware, Standard>

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Client<DynConnector, AwsMiddleware, Standard>

Creates a new client from the service Config.